+++ src/main/java/g1901_2000/s1916_count_ways_to_build_rooms_in_an_ant_colony/Solution.java
@@ -46,7 +46,7 @@ public class Solution {
         long com = 1;
         for (long[] p : list) {
             long choose = c(cnt, (int) (p[0]));
-            cnt -= p[0];
+            cnt -= (int) p[0];
             com = com * choose;
             com %= MOD;
             com = com * p[1];
